feat(ace-b36ts): split deterministic tests into fast and feat suites
Move the package's deterministic coverage into `test/fast` and `test/feat`, add local `ace-test` aliases for the new taxonomy, refresh the layout demo tape, and bump the gem to `0.14.3`.
