+++ ace-b36ts/test/fast/atoms/format_specs_test.rb
@@ -1,6 +1,6 @@
 # frozen_string_literal: true
 
-require_relative "../test_helper"
+require_relative "../../test_helper"
 
 module Ace
   module B36ts
