+++ ace-b36ts/test/fast/commands/cli_test.rb
@@ -1,6 +1,6 @@
 # frozen_string_literal: true
 
-require_relative "../test_helper"
+require_relative "../../test_helper"
 require "ace/support/cli"
 
 class CliTest < AceB36tsTestCase
