+++ ace-b36ts/test/feat/cli_contract_test.rb
